#6879 - move health conditions from ClinicalCourse to Cases backend & vaadin part
