+++ sormas-api/src/main/java/de/symeda/sormas/api/caze/CaseExportDto.java
@@ -29,7 +29,6 @@ import org.apache.commons.lang3.StringUtils;
 
 import de.symeda.sormas.api.CountryHelper;
 import de.symeda.sormas.api.Disease;
-import de.symeda.sormas.api.clinicalcourse.ClinicalCourseDto;
 import de.symeda.sormas.api.clinicalcourse.HealthConditionsDto;
 import de.symeda.sormas.api.contact.FollowUpStatus;
 import de.symeda.sormas.api.contact.QuarantineType;
@@ -1902,7 +1901,7 @@ public class CaseExportDto implements Serializable {
 	@Order(122)
 	@ExportTarget(caseExportTypes = {
 		CaseExportType.CASE_MANAGEMENT })
-	@ExportProperty(ClinicalCourseDto.HEALTH_CONDITIONS)
+	@ExportProperty(CaseDataDto.HEALTH_CONDITIONS)
 	@ExportGroup(ExportGroupType.CASE_MANAGEMENT)
 	public HealthConditionsDto getHealthConditions() {
 		return healthConditions;
