+++ sormas-api/src/main/java/de/symeda/sormas/api/clinicalcourse/ClinicalCourseDto.java
@@ -1,36 +1,16 @@
 package de.symeda.sormas.api.clinicalcourse;
 
-import javax.validation.Valid;
-
 import de.symeda.sormas.api.EntityDto;
 import de.symeda.sormas.api.utils.DataHelper;
 
 public class ClinicalCourseDto extends EntityDto {
 
+	public static final String I18N_PREFIX = "ClinicalCourse";
 	private static final long serialVersionUID = -2664896907352864261L;
 
-	public static final String I18N_PREFIX = "ClinicalCourse";
-
-	public static final String HEALTH_CONDITIONS = "healthConditions";
-
-	@Valid
-	private HealthConditionsDto healthConditions;
-
 	public static ClinicalCourseDto build() {
-		return build(null);
-	}
-
-	public static ClinicalCourseDto build(HealthConditionsDto healthConditions) {
-
 		ClinicalCourseDto clinicalCourse = new ClinicalCourseDto();
 		clinicalCourse.setUuid(DataHelper.createUuid());
-
-		if (healthConditions == null) {
-			clinicalCourse.setHealthConditions(HealthConditionsDto.build());
-		} else {
-			clinicalCourse.setHealthConditions(healthConditions);
-		}
-
 		return clinicalCourse;
 	}
 
@@ -38,11 +18,4 @@ public class ClinicalCourseDto extends EntityDto {
 		return new ClinicalCourseReferenceDto(getUuid());
 	}
 
-	public HealthConditionsDto getHealthConditions() {
-		return healthConditions;
-	}
-
-	public void setHealthConditions(HealthConditionsDto healthConditions) {
-		this.healthConditions = healthConditions;
-	}
 }
