+++ sormas-backend/src/main/java/de/symeda/sormas/backend/caze/CaseService.java
@@ -237,8 +237,7 @@ public class CaseService extends AbstractCoreAdoService<Case> {
 		Root<Case> from = cq.from(getElementClass());
 		from.fetch(Case.SYMPTOMS);
 		from.fetch(Case.THERAPY);
-		Fetch<Case, ClinicalCourse> clinicalCourseFetch = from.fetch(Case.CLINICAL_COURSE);
-		clinicalCourseFetch.fetch(ClinicalCourse.HEALTH_CONDITIONS);
+		from.fetch(Case.HEALTH_CONDITIONS);
 		from.fetch(Case.HOSPITALIZATION);
 		from.fetch(Case.EPI_DATA);
 		from.fetch(Case.PORT_HEALTH_INFO);
@@ -996,7 +995,7 @@ public class CaseService extends AbstractCoreAdoService<Case> {
 
 		filterBuilder = filterBuilder.add(casePath, Case.THERAPY)
 			.add(clinicalCourse)
-			.add(clinicalCourse, ClinicalCourse.HEALTH_CONDITIONS)
+			.add(casePath, Case.HEALTH_CONDITIONS)
 			.add(casePath, Case.MATERNAL_HISTORY)
 			.add(casePath, Case.PORT_HEALTH_INFO)
 			.add(casePath, Case.SORMAS_TO_SORMAS_ORIGIN_INFO)
