+++ sormas-ui/src/main/java/de/symeda/sormas/ui/caze/CaseController.java
@@ -657,7 +657,7 @@ public class CaseController {
 
 					dto.getSymptoms().setOnsetDate(createForm.getOnsetDate());
 					dto.getSymptoms().setUuid(DataHelper.createUuid());
-					dto.getClinicalCourse().getHealthConditions().setUuid(DataHelper.createUuid());
+					dto.getHealthConditions().setUuid(DataHelper.createUuid());
 					dto.getEpiData().setUuid(DataHelper.createUuid());
 					dto.getEpiData().getExposures().forEach(exposure -> {
 						exposure.setUuid(DataHelper.createUuid());
